MNT: reorganize benchmark stuff
