--- benchmarks/__init__.py
@@ -1,3 +1,0 @@
-from . import util  # noqa
-from .ioc_handler import IocHandler  # noqa
-from . import benchmarks  # noqa
