+++ caproto/benchmarking/__init__.py
@@ -1,0 +1,1 @@
+from .util import *  # noqa
